client/requirements.txt
boto3
pandas
mysql-connector-python
flask
requests
uvicorn
uuid